Created groups for the layer control structure
