+++ app/assets/javascripts/viewer-main.js
@@ -6,27 +6,29 @@ require([
 	'dojo/_base/window',
 	'dojo/query',
 	'dojo/dom-attr',
+	'dojo/dom-construct',
+	'dojo/dom-style',
 
 	'dojo/NodeList-traverse',
 	'dojo/domReady!'
-   	], function(dom, ioQuery, on, win, query, domAttr) {
+	], function(dom, ioQuery, on, win, query, domAttr, domConstruct, domStyle) {
 		var origin = [-285401.920, 903401.920];
 		var resolutions = [3440.64, 1720.32, 860.16, 430.08, 215.04, 107.52, 53.76, 26.88, 13.44, 6.72, 3.36, 1.68, 0.84, 0.42, 0.21];
-	    var extent = [-285401.92, 22598.08, 595401.9199999999, 903401.9199999999];
-	    var matrixIds0 = [];
-	    
-	    var i;
-	    
-	    for (i = 0; i < 15; i++) {
-	    	matrixIds0[i] = 'EPSG:28992:' + i;
-	    }
-	    
-	    var tileGrid0 = new ol.tilegrid.WMTS({
+		var extent = [-285401.92, 22598.08, 595401.9199999999, 903401.9199999999];
+		var matrixIds0 = [];
+		
+		var i;
+		
+		for (i = 0; i < 15; i++) {
+			matrixIds0[i] = 'EPSG:28992:' + i;
+		}
+		
+		var tileGrid0 = new ol.tilegrid.WMTS({
 			origin: origin,
 			resolutions: resolutions,
 			matrixIds: matrixIds0
-        });
-	    
+		});
+		
 	    var matrixIds1 = [];
 	    for (i = 0; i < 15; i++) {
 	    	matrixIds1[i] = (i < 10 ? '0' : '') + i;
@@ -232,7 +234,124 @@ require([
 	       		source: projOvrsSource
 	    	});
 	       	
-	       	
+	       	var viewsContainer = dom.byId('views-container');
+			
+			var svrLayerView = dom.byId('svr-layer-view');
+			var svrLayerControl = dom.byId('svr-layer-control');
+			var serviceExpand = on(win.doc, '.js-service-link:click', function(e) {
+				var serviceName = domAttr.get(dojo.query(this).children('.js-link-label')[0], 'innerHTML');
+				var className = domAttr.get(this, 'class');
+				
+				var groupList = dojo.query(this).siblings('.js-insert-groups');
+				
+				if(className == 'js-service-link') {
+					var ul = domConstruct.create('ul', {'id': 'js-ul-groups-' + serviceName});
+					domConstruct.place(ul, this, 'after');
+					
+					for(var i = 0; i < groupList.length; ++i) {
+						var element = dojo.query(groupList)[i];
+						
+						var li = domConstruct.create('li');
+						domAttr.set(li, 'class', 'js-group');
+						var span = domConstruct.create('span');
+						domAttr.set(span, 'class', 'js-group-link');
+						var spanIcon = domConstruct.create('span');
+						domAttr.set(spanIcon, 'class', 'glyphicon glyphicon-plus-sign js-link-icon');
+						var spanLabel = domConstruct.create('span');
+						domAttr.set(spanLabel, 'data-group-id', element.dataset.groupId);
+						domAttr.set(spanLabel, 'innerHTML', ' ' + element.dataset.groupName);
+						domAttr.set(spanLabel, 'class', 'js-link-label');
+						
+						domConstruct.place(li, ul, 'last');
+						domConstruct.place(span, li, 'last');
+						domConstruct.place(spanIcon, span, 'last');
+						domConstruct.place(spanLabel, span, 'last');
+					}
+					
+					domAttr.set(this, 'class', 'js-service-link expanded');
+				}
+				
+				if(className == 'js-service-link expanded') {
+					domAttr.set(this, 'class', 'js-service-link');
+					var groupSetToDel = dojo.query(this).next()[0];
+					domConstruct.destroy(groupSetToDel);
+				}
+			});
+			
+			var groupExpand = on(win.doc, '.js-group-link:click', function(e) {
+				var groupName= domAttr.get(dojo.query(this).children('.js-link-label')[0], 'innerHTML');
+				var className = domAttr.get(this, 'class');
+				
+				if(className == 'js-group-link') {
+					var ul = domConstruct.create('ul', {'id': 'js-ul-layers-' + groupName});
+					domConstruct.place(ul, this, 'after');
+					
+					
+					for(var i = 1; i < 11; ++i) {
+						var li = domConstruct.create('li');
+						domAttr.set(li, 'class', 'js-layer');
+						
+						var span = domConstruct.create('span');
+						domAttr.set(span, 'class', 'js-layer-link');
+						
+						var inputCheckboxLabel = domConstruct.create('label');
+						domAttr.set(inputCheckboxLabel, 'class', 'checkbox-inline');
+						
+						var inputCheckboxInput = domConstruct.create('input');
+						domAttr.set(inputCheckboxInput, 'type', 'checkbox');
+						domStyle.set(inputCheckboxInput, 'position', 'static');
+						
+						var spanLabel = domConstruct.create('span');
+						if(i == 1) {
+							domAttr.set(spanLabel, 'innerHTML', ' Bebouwde kommen in Overijssel');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check bebKomInOvrs');
+						} else if(i == 2) {
+							domAttr.set(spanLabel, 'innerHTML', ' Bebouwde kommen rondom Overijssel');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check bebKomRondOvrs');
+						} else if(i == 3) {
+							domAttr.set(spanLabel, 'innerHTML', ' Bodemgebruik in Overijssel (1993)');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check bodem1993');
+						} else if(i == 4) {
+							domAttr.set(spanLabel, 'innerHTML', ' Bodemgebruik in Overijssel (1996)');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check bodem1996');
+						} else if(i == 5) {
+							domAttr.set(spanLabel, 'innerHTML', ' Gebiedskenmerken stedelijke laag');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check gebStedLaag');
+						} else if(i == 6) {
+							domAttr.set(spanLabel, 'innerHTML', ' Grens projectgebied Vecht Regge');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check grensRegge');
+						} else if(i == 7) {
+							domAttr.set(spanLabel, 'innerHTML', ' Grenzen waterschappen in Overijssel (lijn)');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check grenzenWatLijn');
+						} else if(i == 8) {
+							domAttr.set(spanLabel, 'innerHTML', ' Grenzen waterschappen in Overijssel (vlak)');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check grenzenWatVlak');
+						} else if(i == 9) {
+							domAttr.set(spanLabel, 'innerHTML', ' Nationale parken Weerribben-Wieden en Sallandse Heuvelrug');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check weerribben');
+						} else if(i == 10) {
+							domAttr.set(spanLabel, 'innerHTML', ' Projecten in Overijssel');
+							domAttr.set(inputCheckboxInput, 'class', 'js-layer-check projOvrs');
+						}
+						
+						domAttr.set(spanLabel, 'class', 'js-link-label-checkbox');
+						
+						domConstruct.place(li, ul, 'last');
+						domConstruct.place(span, li, 'last');
+						domConstruct.place(inputCheckboxLabel, span, 'last');
+						domConstruct.place(inputCheckboxInput, inputCheckboxLabel, 'last');
+						domConstruct.place(spanLabel, span, 'last');
+					}
+					
+					domAttr.set(this, 'class', 'js-group-link expanded');
+				}
+				
+				if(className == 'js-group-link expanded') {
+					domAttr.set(this, 'class', 'js-group-link');
+					var layerSetToDel = dojo.query(this).next()[0];
+					domConstruct.destroy(layerSetToDel);
+				}
+			});
 			
 	       	var layerCheck = on(win.doc, '.js-layer-check:change', function(e) {
 				var classNameElement = domAttr.get(this, 'class');
