+++ app/assets/stylesheets/main.less
@@ -11,7 +11,8 @@
 .service-expand {border:none;}
 .service-expand:hover {background-color:transparent;}
 
-.js-service-link, .js-group-link, .js-layer-link {margin-left:10px;cursor: pointer;line-height:15px;}
+.js-service-link, .js-group-link, .js-layer-link {margin-left:10px;cursor: pointer;line-height:15px;color:#000000;}
+.js-service-link:hover, .js-group-link:hover, .js-layer-link:hover {color:#000000;text-decoration:none;}
 .js-link-icon {position:relative;top:2px;}
 .js-link-label-checkbox {position:relative;top:1px;}
