+++ app/views/main.scala.html
@@ -68,7 +68,6 @@
         <script src="@routes.Assets.versioned("lib/bootstrap/js/bootstrap.js")" type="text/javascript"></script>
         <script src="@routes.Assets.versioned("lib/openlayers/ol.js")" type="text/javascript"></script>
 		<script src="@routes.Assets.versioned("lib/dojo/dojo/dojo.js")" type="text/javascript"></script>
-		<script src="@routes.Assets.versioned("javascripts/views.js")" type="text/javascript"></script>
 		<script src="@jsMain" type="text/javascript"></script>
     </body>    
 </html>
